Convert base class to utils file
